requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
httpx>=0.28.0
python-telegram-bot>=22.0
python-dotenv>=1.0.0
fpdf2>=2.8.0
Pillow>=10.0.0
sqlalchemy>=2.0.0
pyyaml>=6.0
aiofiles>=24.0.0
